--- app/src/main/java/com/practicum/playlistmaker/data/NetworkClient.kt
@@ -1,7 +1,0 @@
-package com.practicum.playlistmaker.data
-
-import com.practicum.playlistmaker.data.dto.Response
-
-interface NetworkClient {
-    fun doRequest(dto: Any): Response
-}
