+++ app/src/main/java/com/practicum/playlistmaker/data/history/db/SharedPreferencesDbClient.kt
@@ -1,0 +1,20 @@
+package com.practicum.playlistmaker.data.history.db
+
+import com.practicum.playlistmaker.data.history.DbClient
+
+class SharedPreferencesDbClient : DbClient {
+
+
+
+
+    override fun readTracks() {
+
+
+
+
+    }
+
+    override fun writeTrack() {
+
+    }
+}
