+++ app/src/main/java/com/practicum/playlistmaker/data/searchTracks/NetworkClient.kt
@@ -1,0 +1,7 @@
+package com.practicum.playlistmaker.data.searchTracks
+
+import com.practicum.playlistmaker.data.searchTracks.dto.Response
+
+interface NetworkClient {
+    fun doRequest(dto: Any): Response
+}
