+++ app/src/main/java/com/practicum/playlistmaker/data/searchTracks/dto/Response.kt
@@ -1,0 +1,5 @@
+package com.practicum.playlistmaker.data.searchTracks.dto
+
+open class Response {
+    var resultCode = 0
+}
