+++ app/src/main/java/com/practicum/playlistmaker/data/searchTracks/dto/TracksSearchRequest.kt
@@ -1,0 +1,3 @@
+package com.practicum.playlistmaker.data.searchTracks.dto
+
+data class TracksSearchRequest(val expression: String)
