+++ app/src/main/java/com/practicum/playlistmaker/data/searchTracks/network/ITunesApiService.kt
@@ -1,6 +1,6 @@
-package com.practicum.playlistmaker.data.network
+package com.practicum.playlistmaker.data.searchTracks.network
 
-import com.practicum.playlistmaker.data.dto.TracksSearchResponse
+import com.practicum.playlistmaker.data.searchTracks.dto.TracksSearchResponse
 import retrofit2.Call
 import retrofit2.http.GET
 import retrofit2.http.Query
