--- app/src/main/java/com/practicum/playlistmaker/domain/api/TracksRepository.kt
@@ -1,12 +1,0 @@
-package com.practicum.playlistmaker.domain.api
-
-import com.practicum.playlistmaker.domain.models.SearchTracksResult
-import com.practicum.playlistmaker.domain.models.Track
-
-interface TracksRepository {
-    fun searchTracks(
-        expression: String
-    ): SearchTracksResult
-
-    fun writeTrack(trackClicked: Track)
-}
