+++ app/src/main/java/com/practicum/playlistmaker/domain/searchTracks/api/TracksInteractor.kt
@@ -1,0 +1,11 @@
+package com.practicum.playlistmaker.domain.searchTracks.api
+
+import com.practicum.playlistmaker.domain.searchTracks.models.SearchTracksResult
+
+interface TracksInteractor {
+    fun searchTracks(expression: String, consumer: TracksConsumer)
+
+    interface TracksConsumer {
+        fun consume(result: SearchTracksResult)
+    }
+}
