+++ app/src/main/java/com/practicum/playlistmaker/domain/searchTracks/api/TracksRepository.kt
@@ -1,0 +1,9 @@
+package com.practicum.playlistmaker.domain.searchTracks.api
+
+import com.practicum.playlistmaker.domain.searchTracks.models.SearchTracksResult
+
+interface TracksRepository {
+    fun searchTracks(
+        expression: String
+    ): SearchTracksResult
+}
