+++ app/src/main/java/com/practicum/playlistmaker/domain/searchTracks/impl/TracksInteractorImpl.kt
@@ -1,9 +1,7 @@
-package com.practicum.playlistmaker.domain.impl
+package com.practicum.playlistmaker.domain.searchTracks.impl
 
-import android.util.Log
-import com.practicum.playlistmaker.domain.api.TracksInteractor
-import com.practicum.playlistmaker.domain.api.TracksRepository
-import com.practicum.playlistmaker.domain.models.Track
+import com.practicum.playlistmaker.domain.searchTracks.api.TracksInteractor
+import com.practicum.playlistmaker.domain.searchTracks.api.TracksRepository
 import java.util.concurrent.Executors
 
 class TracksInteractorImpl(private val repository: TracksRepository): TracksInteractor {
@@ -15,9 +13,5 @@ class TracksInteractorImpl(private val repository: TracksRepository): TracksInte
         executor.execute {
             consumer.consume(repository.searchTracks(expression))
         }
-    }
-
-    override fun writeHistoryTracks(trackClicked: Track) {
-
     }
 }
