+++ app/src/main/java/com/practicum/playlistmaker/domain/searchTracks/models/PlayerStatus.kt
@@ -1,4 +1,4 @@
-package com.practicum.playlistmaker.domain.models
+package com.practicum.playlistmaker.domain.searchTracks.models
 
 enum class PlayerStatus {
     STATE_DEFAULT,
